All changes added at once because I didnt fork this before I started changing files (sorry me, or next developer!).
It's in the author of UM's best interest to not change the structure of
the plugin too much, so it should be fine to merge this repo with the
new version. We can do a test-run of the merge to see. If a manual merge
is required, its not a big deal, because the changes are mainly
condensed blocks.

* 3 new rows in admin
* the necessary ui adjustments to accomodate the 3 new variations
* print button for subscribers and editors